Update README.md
PacMan Motion and Edge Detection Simulation Project

Description: The aim of this project is to demonstrate ability to apply certain JavaScript skills such as use of functions. In this project, I applied such functions such as checkPageBounds, setIntervals, and setTimeouts to make PacMan move from one side of the screen to the other then reversing upon reaching the left and right edges of the screen.
Usage: This JavaScript exercise is quite beneficial and can be useful in programs or apps dealing with animations, motion, edge detection, and timing.  
Credit / Support: This exercise was completed using several resources such as MIT Xpro bootcamp, W3 Schools, and MDN.